use blitz rpc for stats page
